Add files via upload
url_Grabber is used to make a .txt file containing all of the .zip files we want to work with.,
url_Sotrage is used to generate the "storage" dataframe that contains all of the relevant information for each zip file: ['variable', 'period', 'year', 'url_list']
variable: the info we are working with. This is wind, temperature, AQI, ozone, etc
period: the period of data colelction: yearly, hourly, daily, etc
year: the year the data was collected
url: the relevant zip files for each row of data. these will be used with the requests library to actually get the csv files when working on the front end.